+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Loading classes using ClassHelper.getDefaultClassLoader().loadClass() 
+instead of Class.forName().
+
 ## [1.3.0] - 2024-04-01
 
 ### Added